Implement react-router-dom. See full commit msg:
You can now navigate between the Home page and the Profile page. I set
it up so that when you do npm start, the localhost you go to will
display 'Home'. And then, add '/profile' to the end of the URL to go to
the Profile page, where it should display 'Profile'.