render logo svg to topbar:
I'm thinking of making a new version of the logo where the icon is on the side and
the lettering "SoundSage" is to the right of it, inline with the icon. Right now,
it is not very readable! Another issue: it's all black so it disappears upon
changing the website theme. Perhaps I could change the logo color to green.