Implement Blog Object and temporary db functionality! See To-Do:
This commit contains implementation for a few things: a blog collection file, a temporary JSON
database file for blogs (we can change later), and a blog_home display feature that shows
ALL of the blogs posted on the main page.
To get the database file working, we need to run TWO terminals: one with `npm start`,
and another with `npm run server`. The latter command should default to port 8000, and if it
doesn't, the program may not run due to port 8000 being (temporarily) hard-coded into the
implementation.
TO-DO: The blog objects should also have a "SONG" member element, but they don't right now.
Perhaps we implement an empty object later for a placeholder?