Add post creation functionality, has major bugs:
After you fill out the form on the "Create" tab, the blogs
are still not visible on the home page because, currently, the
blog demos are hard-coded and later we will have it retrieve from
the backend. Also, I added a new BlogPage file which is where you'd
get to after clicking on a blog preview on the home page. But this
also isn't working yet and likely will work once we retrieve data from
backend in the location where we hard-coded the demo blog.